requirements.txt
Django==2.1.3
django-debug-toolbar==1.10.1
python-decouple==3.1
six==1.11.0
sqlparse==0.2.4
